refactor: refactor clip board
